Add avx2 for forward txfms
