+++ src/asm/x86/transform/mod.rs
@@ -7,4 +7,5 @@
 // Media Patent License 1.0 was not distributed with this source code in the
 // PATENTS file, you can obtain it at www.aomedia.org/license/patent.
 
+pub mod forward;
 pub mod inverse;
