+++ src/lib.rs
@@ -77,6 +77,9 @@ mod hawktracer {
 #[cfg(any(cargo_c, feature = "capi"))]
 pub mod capi;
 
+#[macro_use]
+mod transform;
+
 mod activity;
 pub(crate) mod asm;
 mod cpu_features;
@@ -87,7 +90,6 @@ mod predict;
 mod quantize;
 mod rdo;
 mod rdo_tables;
-mod transform;
 #[macro_use]
 mod util;
 mod cdef;
